+++ Bjorn/docs/release-notes.md
@@ -6,6 +6,7 @@
 
 * Fixed the Lil' Hadron Collider card analyzing non-analyzable cards.
 * Fixed the Lil' Hadron Collider card not counting as analyzing cards for some effects (for example, the Side Projects artifact).
+* Fixed "On Analyze" action trigger order.
 
 ## 1.0.0
 Released 29 March 2025.